+++ components/Layout/Layout.module.scss
@@ -1,4 +1,4 @@
-@use 'styles/variables' as *;
+@use '../../styles/variables' as *;
 
 .contentWrapper {
   background-color: $white;
