+++ pages/_app.tsx
@@ -1,5 +1,5 @@
 import type { AppProps } from 'next/app';
-import Layout from '../components/main/Layout';
+import Layout from '../components/Layout';
 import '../styles/main.scss';
 
 function MyApp({ Component, pageProps }: AppProps) {
