+++ pages/index.js
@@ -1,2 +1,2 @@
-import Home from '../components/page-components/Home';
+import Home from '../components/Home';
 export default Home;
