+++ styles/_globals.scss
@@ -33,7 +33,7 @@ a {
   padding: 10px 40px;
   border-radius: 10px;
   letter-spacing: 2px;
-  border: 2px solid $gray;
+  border: 2px solid $borderGray;
   background-color: transparent;
   text-transform: uppercase;
   font-weight: 700;
